+++ NEWS.md
@@ -16,6 +16,7 @@
   - For convenice, named or numbered types or collections in the comicbox schema
     can also parse by their simple name instead of requiring a complex object.
   - Comictagger schema supports series_aliases and title_aliases as reprints
+  - CB7 archive read support. Comic archives in 7zip archives.
 
 - Fixes
   - ComicInfo.xml StoryArcs tag was not loaded.